Added clustering for all cross sections.
Reorders the codebook to make cluster 1 come first (in energy), 2 next, etc.

Clusters RRR for all cross sections if their threshold is not above the RRR.

Thins the energy grid to combine contiguous ranges that share the same cluster.

Updates the threshold indexes for each reaction to be consistent with the thinned grid.

Removed a log averaging in defining the average cross sections so sum of partials stays equal to total.

Still need to oversee the creation of grid_index with correct cluster indexing.