Differentiated between energy grid and xs grid
The grid_index of a nuclide now always holds the energy grid index.
If no clustering is used, the same index is used for xs and energy for
a given nuclide. If clustering is used, the mapping depends on the particle
energy and may require a lookup in the codebook, which maps energy to cluster.